Refactor authentication and user management
- Updated AuthApplication to use @ConfigurationPropertiesScan instead of @EnableConfigurationProperties.
- Refactored SecurityConfig to replace CustomAuthenticationEntryPoint with JwtAuthenticationEntryPoint.
- Added UserController to handle user profile retrieval.
- Changed UserPrincipal's customerId type from int to long.
- Modified ApiResponse to use ErrorResponse instead of Object for error handling.
- Introduced ErrorResponse class to standardize error responses.
- Created GetUserResponse for user profile responses.
- Added ErrorCodeEnum for consistent error codes across the application.
- Removed AuthException and replaced it with AuthServiceException for better error handling.
- Implemented JwtAccessDeniedHandler and JwtAuthenticationEntryPoint for JWT-based security.
- Refactored JwtManager to support access and refresh tokens with improved expiration handling.
- Updated application configuration files to use YAML format and environment variables for sensitive data.
- Created UserService and UserServiceImpl for user profile management.

+++ src/main/java/com/putraawali/auth/dto/request/UserPrincipal.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class UserPrincipal {
     private String email;
-    private int customerId;
+    private long customerId;
 }
